ut5.0
Intune Enroll for Windows devices allows semi-automatic enrollment in Intune when an Azure P1 license is not available. This avoids the need for manual enrollment on each individual device and temporarily makes every user a local administrator to perform the enrollment.
The script can be executed via GPO or distribution software and can be associated to run with a batch file composed as follows:
Powershell.exe -ExecutionPolicy Bypass -ExecutionPolicy RemoteSigned .\ut5.0
The script will check on target devices for the presence of a physical user connected to the PC and verify the presence of its flag file to avoid re-execution on devices that have already received it or have enrolled in other ways through a registry key.
If enrollment has not already been done, the script will proceed to create the c:\enroll folder and its task and file structure, and then execute.
It will add the current user to the administrators group and prompt for a restart; upon user login, it will display the enrollment screen, which will auto-fill if the device is already registered in Azure AD in Hybrid Join, otherwise the user will need to enter their credentials.
Once enrollment is complete, the control script will remove the user or users added to the administrators group, leaving intact the users who already belong to it, remove the scripts and tasks, and create the device.enroll flag file that will block further execution of the script.
The script will generate log files for hostname.txt and user.txt.
If the SFTP part is configured, the script will upload the device.txt log to the FTP within the tenant name folder and update it with the progress status every 30 minutes.